- Update of Bullet class: v0.2 - Bullets aim at mouse position: Bullets only aimable in "bulletspeed"-number of directions per quadrant (if 0 > velX < 1, actual velX -> 0 because x is int!!) => Coordinates of GameObject.java changed from int to float => smooth aiming at mouse cursor
TODO
(on pause) Implementation of player position
(optional) Changing mouse cursor to crosshair while in Window

+++ src/Woralcoholics/game/Enemy.java
@@ -64,7 +64,7 @@ public class Enemy extends GameObject {
     //Enemy is displaying in this Color for now
     public void render(Graphics g) {
         g.setColor(Color.YELLOW);
-        g.fillRect(x, y, 32, 32);
+        g.fillRect((int)x, (int)y, 32, 32);
 
         /* -> Just for testing Bounding Boxes around the Enemys
         Graphics2D g2d = (Graphics2D) g;
@@ -75,10 +75,10 @@ public class Enemy extends GameObject {
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 32, 32);
+        return new Rectangle((int)x, (int)y, 32, 32);
     }
 
     public Rectangle getBoundsAround() {
-        return new Rectangle(x - 16, y - 16, 64, 64);
+        return new Rectangle((int)x - 16, (int)y - 16, 64, 64);
     }
 }
